refactoring db types
