+++ src/app/amenities/add/addamenityform.tsx
@@ -1,21 +1,10 @@
 import React from 'react'
-import { useProfileContext } from '@/context/ProfileContext';
-import Spinner from '@/components/spinner/Spinner';
-import { redirect } from 'next/navigation';
 
 export default function AddAmenityForm() {
-    const { profile, loading, error, authState } = useProfileContext();
-
-    if (loading) return <Spinner />;
-
-    if (authState.status === 'SIGNED_OUT' || null) {
-        redirect('/')
-    }
 
     return (
         <>
-            <div>
-                <h2>Welcome, {profile.full_name || 'Guest!'}</h2>
+            <div>                
                 <div className="min-h-screen flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8">
                     <div className="max-w-md w-full space-y-8 p-10 bg-white rounded-xl shadow-lg">
                         <div className="grid  gap-8 grid-cols-1">
